requirements.txt
amqp==2.5.1
billiard==3.6.1.0
boto3==1.9.228
botocore==1.12.228
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-smtp-ssl==1.0
django-storages==1.7.2
djangorestframework==3.10.3
djangorestframework-stubs==1.0.0
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.22
jmespath==0.9.4
kombu==4.6.4
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.2
redis==3.3.8
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
whitenoise==4.1.3
zipp==0.6.0
